use size_t instead of uint8_t for hint ids

+++ soh/soh/Enhancements/randomizer/3drando/hints.cpp
@@ -40,7 +40,7 @@ const CustomMessage& HintText::GetObscure() const {
     return obscureText.size() > 0 ? RandomElement(obscureText) : clearText;
 }
 
-const CustomMessage& HintText::GetObscure(uint8_t selection) const {
+const CustomMessage& HintText::GetObscure(size_t selection) const {
     if (obscureText.size() > selection) {
         return obscureText[selection];
     } else if (obscureText.size() > 0) {
@@ -53,7 +53,7 @@ const CustomMessage& HintText::GetAmbiguous() const {
     return ambiguousText.size() > 0 ? RandomElement(ambiguousText) : clearText;
 }
 
-const CustomMessage& HintText::GetAmbiguous(uint8_t selection) const {
+const CustomMessage& HintText::GetAmbiguous(size_t selection) const {
     if (ambiguousText.size() > selection) {
         return ambiguousText[selection];
     } else if (ambiguousText.size() > 0) {
@@ -62,15 +62,15 @@ const CustomMessage& HintText::GetAmbiguous(uint8_t selection) const {
     return clearText;
 }
 
-uint8_t HintText::GetAmbiguousSize() const {
-    return static_cast<uint8_t>(ambiguousText.size());
+size_t HintText::GetAmbiguousSize() const {
+    return ambiguousText.size();
 }
 
-uint8_t HintText::GetObscureSize() const {
-    return static_cast<uint8_t>(obscureText.size());
+size_t HintText::GetObscureSize() const {
+    return obscureText.size();
 }
 
-const CustomMessage& HintText::GetHintMessage(uint8_t selection) const {
+const CustomMessage& HintText::GetHintMessage(size_t selection) const {
     auto ctx = Rando::Context::GetInstance();
     if (ctx->GetOption(RSK_HINT_CLARITY).Is(RO_HINT_CLARITY_OBSCURE)) {
         return GetObscure(selection);
